Now using "Last Updated" date and UTC for commit age comparison
I found multiple instances where the "Build Date" in `pacman -Si` is
before the date of most recent relevant git commit for that version of
the package. My best guess is that maintainers are building the package
first, then pushing the git commit. To combat this, I switched back to
using the "Last Updated" date from the packages.archlinux.org page. I had
to set this time to 1 second before the next day to get the comparison
right. I also made sure everything is in UTC time.